Provide more information about the task status
