+++ src/HotChocolate/Core/src/Abstractions/Execution/Tasks/ParallelExecutionTask.cs
@@ -11,6 +11,7 @@ namespace HotChocolate.Execution
     /// </summary>
     public abstract class ParallelExecutionTask : IExecutionTask
     {
+        private ExecutionTaskStatus _completionStatus = ExecutionTaskStatus.Completed;
         private Task? _task;
 
         /// <summary>
@@ -22,7 +23,7 @@ namespace HotChocolate.Execution
         public ExecutionTaskKind Kind => ExecutionTaskKind.Parallel;
 
         /// <inheritdoc />
-        public bool IsCompleted => _task?.IsCompleted ?? false;
+        public ExecutionTaskStatus Status { get; private set; }
 
         /// <inheritdoc />
         public IExecutionTask? Next { get; set; }
@@ -41,7 +42,10 @@ namespace HotChocolate.Execution
 
         /// <inheritdoc />
         public void BeginExecute(CancellationToken cancellationToken)
-            => _task = ExecuteInternalAsync(cancellationToken).AsTask();
+        {
+            Status = ExecutionTaskStatus.Running;
+            _task = ExecuteInternalAsync(cancellationToken).AsTask();
+        }
 
         /// <inheritdoc />
         public Task WaitForCompletionAsync(CancellationToken cancellationToken)
@@ -55,14 +59,20 @@ namespace HotChocolate.Execution
                 {
                     await ExecuteAsync(cancellationToken).ConfigureAwait(false);
                 }
+
+                Status = _completionStatus;
             }
             catch (OperationCanceledException)
             {
+                Status = ExecutionTaskStatus.Faulted;
+
                 // If we run into this exception the request was aborted.
                 // In this case we do nothing and just return.
             }
             catch (Exception ex)
             {
+                Status = ExecutionTaskStatus.Faulted;
+
                 if (cancellationToken.IsCancellationRequested)
                 {
                     // if cancellation is request we do no longer report errors to the
@@ -87,6 +97,14 @@ namespace HotChocolate.Execution
         protected abstract ValueTask ExecuteAsync(CancellationToken cancellationToken);
 
         /// <summary>
+        /// Completes the task as faulted.
+        /// </summary>
+        protected void Faulted()
+        {
+            _completionStatus = ExecutionTaskStatus.Faulted;
+        }
+
+        /// <summary>
         /// Resets the state of this task in case the task object is reused.
         /// </summary>
         protected void Reset()
@@ -97,6 +115,8 @@ namespace HotChocolate.Execution
             State = null;
             IsSerial = false;
             IsRegistered = false;
+            _completionStatus = ExecutionTaskStatus.Completed;
+            Status = ExecutionTaskStatus.WaitingToRun;
         }
     }
 }
