+++ src/HotChocolate/Core/src/Execution/Processing/Tasks/ResolverTask.cs
@@ -1,16 +1,22 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
+using System.Collections.Immutable;
 using System.Threading;
 using System.Threading.Tasks;
+using HotChocolate.Execution.Instrumentation;
 using HotChocolate.Types;
 using Microsoft.Extensions.ObjectPool;
 
 namespace HotChocolate.Execution.Processing.Tasks
 {
-    internal sealed class ResolverTask : ResolverTaskBase
+    internal sealed partial class ResolverTask : IExecutionTask
     {
         private readonly ObjectPool<ResolverTask> _objectPool;
+        private readonly MiddlewareContext _resolverContext = new();
+        private readonly List<IExecutionTask> _taskBuffer = new();
+        private IOperationContext _operationContext = default!;
+        private ISelection _selection = default!;
+        private ExecutionTaskStatus _completionStatus = ExecutionTaskStatus.Completed;
         private Task? _task;
 
         public ResolverTask(ObjectPool<ResolverTask> objectPool)
@@ -18,231 +24,193 @@ namespace HotChocolate.Execution.Processing.Tasks
             _objectPool = objectPool ?? throw new ArgumentNullException(nameof(objectPool));
         }
 
-        public override ExecutionTaskKind Kind => ExecutionTaskKind.Parallel;
+        /// <summary>
+        /// Gets access to the resolver context for this task.
+        /// </summary>
+        internal MiddlewareContext ResolverContext => _resolverContext;
 
-        public override void BeginExecute(CancellationToken cancellationToken)
+        /// <summary>
+        /// Gets access to the operation context.
+        /// </summary>
+        private IOperationContext OperationContext => _operationContext;
+
+        /// <summary>
+        /// Gets access to the diagnostic events.
+        /// </summary>
+        private IExecutionDiagnosticEvents DiagnosticEvents => OperationContext.DiagnosticEvents;
+
+        /// <summary>
+        /// Gets the selection for which a resolver is executed.
+        /// </summary>
+        internal ISelection Selection => _selection;
+
+        /// <inheritdoc />
+        public ExecutionTaskKind Kind => ExecutionTaskKind.Parallel;
+
+        /// <inheritdoc />
+        public ExecutionTaskStatus Status { get; private set; }
+
+        /// <inheritdoc />
+        public IExecutionTask? Next { get; set; }
+
+        /// <inheritdoc />
+        public IExecutionTask? Previous { get; set; }
+
+        /// <summary>
+        /// Gets access to the internal result map into which the task will write the result.
+        /// </summary>
+        public ResultMap ResultMap { get; private set; } = default!;
+
+        /// <summary>
+        /// Gets the completed value of this task.
+        /// </summary>
+        public object? CompletedValue { get; private set; }
+
+        /// <inheritdoc />
+        public object? State { get; set; }
+
+        /// <inheritdoc />
+        public bool IsSerial { get; set; }
+
+        /// <inheritdoc />
+        public bool IsRegistered { get; set; }
+
+        /// <inheritdoc />
+        public void BeginExecute(CancellationToken cancellationToken)
         {
+            Status = ExecutionTaskStatus.Running;
             _task = ExecuteAsync(cancellationToken);
         }
 
-        public override Task WaitForCompletionAsync(CancellationToken cancellationToken) =>
+        /// <inheritdoc />
+        public Task WaitForCompletionAsync(CancellationToken cancellationToken) =>
             _task ?? Task.CompletedTask;
 
-        private async Task ExecuteAsync(CancellationToken cancellationToken)
+        /// <summary>
+        /// Completes the resolver result.
+        /// </summary>
+        /// <param name="success">Defines if the resolver succeeded without errors.</param>
+        /// <param name="cancellationToken">The execution cancellation token.</param>
+        private void CompleteValue(bool success, CancellationToken cancellationToken)
         {
+            object? completedValue = null;
+
             try
             {
-                using (DiagnosticEvents.ResolveFieldValue(ResolverContext))
+                // we will only try to complete the resolver value if there are no known errors.
+                if (success)
                 {
-                    var success = await TryExecuteAsync(cancellationToken).ConfigureAwait(false);
-                    CompleteValue(success, cancellationToken);
+                    if (ValueCompletion.TryComplete(
+                        _operationContext,
+                        _resolverContext,
+                        (ISelection)_resolverContext.Selection,
+                        _resolverContext.Path,
+                        _selection.Type,
+                        _resolverContext.ResponseName,
+                        _resolverContext.ResponseIndex,
+                        _resolverContext.Result,
+                        _taskBuffer,
+                        out completedValue) &&
+                        _selection.TypeKind is not TypeKind.Scalar and not TypeKind.Enum &&
+                        completedValue is IHasResultDataParent result)
+                    {
+                        result.Parent = _resolverContext.ResultMap;
+                    }
+
+                    if (_taskBuffer.Count > 0)
+                    {
+                        _operationContext.Scheduler.Register(_taskBuffer);
+                    }
                 }
             }
-            catch
+            catch (OperationCanceledException)
             {
-                // we suppress any exception if the cancellation was requested.
-                if (!cancellationToken.IsCancellationRequested)
-                {
-                    throw;
-                }
-            }
-            finally
-            {
-                IsCompleted = true;
-                OperationContext.Scheduler.Complete(this);
-                _objectPool.Return(this);
-            }
-        }
-
-        private async ValueTask<bool> TryExecuteAsync(CancellationToken cancellationToken)
-        {
-            try
-            {
-                if (cancellationToken.IsCancellationRequested)
-                {
-                    return false;
-                }
-
-                if (Selection.Arguments.IsFinalNoErrors)
-                {
-                    ResolverContext.Arguments = Selection.Arguments;
-                    await ExecuteResolverPipelineAsync(cancellationToken).ConfigureAwait(false);
-                    return true;
-                }
-
-                if (Selection.Arguments.TryCoerceArguments(
-                    ResolverContext,
-                    out IReadOnlyDictionary<NameString, ArgumentValue>? coercedArgs))
-                {
-                    ResolverContext.Arguments = coercedArgs;
-                    await ExecuteResolverPipelineAsync(cancellationToken).ConfigureAwait(false);
-                    return true;
-                }
+                // If we run into this exception the request was aborted.
+                // In this case we do nothing and just return.
+                _completionStatus = ExecutionTaskStatus.Faulted;
+                return;
             }
             catch (Exception ex)
             {
-                if (!cancellationToken.IsCancellationRequested)
+                if (cancellationToken.IsCancellationRequested)
                 {
-                    ResolverContext.ReportError(ex);
-                    ResolverContext.Result = null;
-                }
-            }
-
-            return false;
-        }
-
-        private async ValueTask ExecuteResolverPipelineAsync(CancellationToken cancellationToken)
-        {
-            await ResolverContext.ResolverPipeline!(ResolverContext).ConfigureAwait(false);
-
-            if (ResolverContext.Result is null)
-            {
-                return;
-            }
-
-            if (ResolverContext.Result is IError error)
-            {
-                ResolverContext.ReportError(error);
-                ResolverContext.Result = null;
-                return;
-            }
-
-            // if we are not a list we do not need any further result processing.
-            if (!Selection.IsList)
-            {
-                return;
-            }
-
-            if (Selection.IsStreamable)
-            {
-                StreamDirective streamDirective =
-                    Selection.SyntaxNode.Directives.GetStreamDirective(
-                        ResolverContext.Variables)!;
-                if (streamDirective.If)
-                {
-                    ResolverContext.Result =
-                        await CreateStreamResultAsync(streamDirective)
-                            .ConfigureAwait(false);
+                    // if cancellation is request we do no longer report errors to the
+                    // operation context.
                     return;
                 }
+
+                _resolverContext.ReportError(ex);
+                _resolverContext.Result = null;
+                _completionStatus = ExecutionTaskStatus.Faulted;
             }
 
-            if (Selection.MaybeStream)
+            CompletedValue = completedValue;
+
+            if (completedValue is null && _resolverContext.Field.Type.IsNonNullType())
             {
-                ResolverContext.Result =
-                    await CreateListFromStreamAsync()
-                        .ConfigureAwait(false);
-                return;
+                // if we detect a non-null violation we will stash it for later.
+                // the non-null propagation is delayed so that we can parallelize better.
+                _completionStatus = ExecutionTaskStatus.Faulted;
+                _operationContext.Result.AddNonNullViolation(
+                    _resolverContext.Selection.SyntaxNode,
+                    _resolverContext.Path,
+                    _resolverContext.ResultMap);
             }
-
-            switch (ResolverContext.Result)
+            else
             {
-                case IExecutable executable:
-                    ResolverContext.Result = await executable
-                        .ToListAsync(cancellationToken)
-                        .ConfigureAwait(false);
-                    break;
-
-                case IQueryable queryable:
-                    ResolverContext.Result = await Task.Run(() =>
-                    {
-                        var items = new List<object?>();
-                        foreach (var o in queryable)
-                        {
-                            items.Add(o);
-
-                            if (cancellationToken.IsCancellationRequested)
-                            {
-                                break;
-                            }
-                        }
-
-                        return items;
-                    }, cancellationToken);
-                    break;
+                _resolverContext.ResultMap.SetValue(
+                    _resolverContext.ResponseIndex,
+                    _resolverContext.ResponseName,
+                    completedValue,
+                    _resolverContext.Field.Type.IsNullableType());
             }
         }
 
-        private async ValueTask<List<object?>> CreateStreamResultAsync(
-            StreamDirective streamDirective)
+        /// <summary>
+        /// Initializes this task after it is retrieved from its pool.
+        /// </summary>
+        public void Initialize(
+            IOperationContext operationContext,
+            ISelection selection,
+            ResultMap resultMap,
+            int responseIndex,
+            object? parent,
+            Path path,
+            IImmutableDictionary<string, object?> scopedContextData)
         {
-            IAsyncEnumerable<object?> enumerable = Selection.CreateStream(ResolverContext.Result!);
-            IAsyncEnumerator<object?> enumerator = enumerable.GetAsyncEnumerator();
-            var next = true;
-
-            try
-            {
-                var list = new List<object?>();
-                var initialCount = streamDirective.InitialCount;
-                var count = 0;
-
-                if (initialCount > 0)
-                {
-                    while(next)
-                    {
-                        count++;
-                        next = await enumerator.MoveNextAsync().ConfigureAwait(false);
-                        list.Add(enumerator.Current);
-
-                        if (count >= initialCount)
-                        {
-                            break;
-                        }
-                    }
-                }
-
-                if (next)
-                {
-                    // if the stream has more items than the initial requested items then we will
-                    // defer the rest of the stream.
-                    OperationContext.Scheduler.DeferredWork.Register(
-                        new DeferredStream(
-                            Selection,
-                            streamDirective.Label,
-                            ResolverContext.Path,
-                            ResolverContext.Parent<object>(),
-                            count - 1,
-                            enumerator,
-                            ResolverContext.ScopedContextData));
-                }
-
-                return list;
-            }
-            finally
-            {
-                if (!next)
-                {
-                    // if there is no deferred work we will just dispose the enumerator.
-                    // in the case we have deferred work, the deferred stream handler is
-                    // responsible of handling the dispose.
-                    await enumerator.DisposeAsync().ConfigureAwait(false);
-                }
-            }
+            _operationContext = operationContext;
+            _selection = selection;
+            _resolverContext.Initialize(
+                operationContext,
+                selection,
+                resultMap,
+                responseIndex,
+                parent,
+                path,
+                scopedContextData);
+            ResultMap = resultMap;
         }
 
-        private async ValueTask<List<object?>> CreateListFromStreamAsync()
+        /// <summary>
+        /// Resets the resolver task before returning it to the pool.
+        /// </summary>
+        /// <returns></returns>
+        public bool Reset()
         {
-            IAsyncEnumerable<object?> enumerable = Selection.CreateStream(ResolverContext.Result!);
-            var list = new List<object?>();
-
-            await foreach (var item in enumerable
-                .WithCancellation(ResolverContext.RequestAborted)
-                .ConfigureAwait(false))
-            {
-                list.Add(item);
-            }
-
-            return list;
-        }
-
-        public void CompleteUnsafe()
-        {
-            if (!IsCompleted)
-            {
-                IsCompleted = true;
-                OperationContext.Scheduler.Complete(this);
-                _objectPool.Return(this);
-            }
+            _completionStatus = ExecutionTaskStatus.Completed;
+            _operationContext = default!;
+            _selection = default!;
+            _resolverContext.Clean();
+            ResultMap = default!;
+            CompletedValue = null;
+            Status = ExecutionTaskStatus.WaitingToRun;
+            IsSerial = false;
+            IsRegistered = false;
+            Next = null;
+            Previous = null;
+            State = null;
+            _taskBuffer.Clear();
+            return true;
         }
     }
 }
